refactored and moved Volume class
