+++ cfme/storage/volume.py
@@ -4,8 +4,7 @@ from functools import partial
 from navmazing import NavigateToSibling
 
 import cfme.fixtures.pytest_selenium as sel
-from cfme.exceptions import DestinationNotFound
-from cfme.web_ui import match_location, InfoBlock, PagedTable, toolbar as tb
+from cfme.web_ui import match_location, PagedTable, toolbar as tb
 from utils.appliance import Navigatable
 from utils.appliance.implementations.ui import CFMENavigateStep, navigator, navigate_to
 
@@ -25,30 +24,16 @@ class Volume(Navigatable):
 @navigator.register(Volume, 'All')
 class All(CFMENavigateStep):
     def prerequisite(self):
-        if self.obj.appliance.version >= '5.7':
-            raise DestinationNotFound('Cloud Volumes does not exist in CFME 5.7+')
         navigate_to(self.obj.appliance.server, 'LoggedIn')
 
     def am_i_here(self):
         return match_volumes(summary='Cloud Volumes')
 
-    def step(self, *args, **kwargs):
-        self.prerequisite_view.navigation.select('Compute', 'Clouds', 'Volumes')
-
-
-@navigator.register(Volume, 'AllByProvider')
-class AllByProvider(CFMENavigateStep):
-    def prerequisite(self):
-        if self.obj.appliance.version >= '5.7':
-            raise DestinationNotFound('Cloud Volumes does not exist in CFME 5.7+')
-        navigate_to(self.obj.appliance.server, 'LoggedIn')
-
-    def am_i_here(self):
-        return match_provider(summary='{} (All Cloud Volumes)'.format(self.obj.provider.name))
-
     def step(self):
-        navigate_to(self.obj.provider, 'Details')
-        sel.click(InfoBlock('Relationships', 'Cloud volumes').element)
+        if self.obj.appliance.version < '5.7':
+            self.prerequisite_view.navigation.select('Compute', 'Clouds', 'Volumes')
+        else:
+            self.prerequisite_view.navigation.select('Storage', 'Volumes')
 
 
 @navigator.register(Volume, 'Details')
@@ -58,7 +43,7 @@ class Details(CFMENavigateStep):
     def am_i_here(self):
         return match_volumes(summary='{} (Summary)'.format(self.obj.name))
 
-    def step(self, *args, **kwargs):
+    def step(self):
         tb.select('List View')
         sel.click(list_tbl.find_row_by_cell_on_all_pages(
             {'Name': self.obj.name, 'Cloud Provider': self.obj.provider.name}))
